UserCache: env var to enable agressive cache writing
When using oca-port as a Python package to collect migration data where
repositories are cloned on a filesystem with very low IO perf, the
process could take a lot of time to finish (every underlying git operations
to retrieve for instance updated file paths of a commit are very
costly). Such process could be killed before it actually finishes (and
write the data in cache).

By setting the OCA_PORT_AGRESSIVE_CACHE_WRITE environment variable, we
force oca-port to update the cache very often, generating more IO during
the first analysis of a module, but making the next process execution
much faster.

As this is an advanced option, it is exposed through an environment
variable to not clutter the usual parameters.